subscribe events/listings
update some code for subscribe button and js code too. But they don’t
really work on current version. But it worked in Meteor example Todos
project. I will continue working on it and any of you know more about
this want to make change feel free to do it.